refactor(backend): decompose quizzes.py into a per-resource package
The 635-line app/api/v1/quizzes.py handled five different responsibilities
(CRUD, student submit, teacher attempt listing, manual grading, extra-
attempts grants) in a single file with tangled helpers. Split it into a
proper package so each file speaks to one concern:

  app/api/v1/quizzes/
    __init__.py    side-effect imports + re-export `router`
    _router.py     shared APIRouter(prefix="/quizzes")
    _deps.py       verify_quiz_owner() local to this package
    crud.py        create / read / update / delete
    attempts.py    /submit, /attempts, /my-attempts
    grading.py     /pending-answers, PATCH /answers/{id}
    extra.py       POST/GET /extra-attempts

Business logic that was embedded in the router now lives in a new
app/services/quiz_service.py:

  - AUTO_GRADED_QUESTION_TYPES, MANUAL_GRADED_QUESTION_TYPES constants
  - ensure_attempts_available()
  - index_quiz_options()
  - grade_auto_answer()
  - persist_answers()
  - upsert_passed_chapter_progress()
  - recompute_attempt_grade()

Public surface is unchanged:
  - Same 12 endpoints, same paths, same response models.
  - `from app.api.v1 import quizzes; quizzes.router` still works
    (app/api/v1/__init__.py is untouched).

All 396 tests pass; ruff/mypy clean on app/.

Made-with: Cursor
